backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.119.0"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.11.0"
httpx = "^0.28.1"
python-dotenv = "^1.1.1"
argon2-cffi = "^25.1.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
alembic = "^1.17.0"
uvicorn = "^0.32.0"
python-dateutil = "^2.9.0.post0"
sqlalchemy-pytds = "^1.0.2"
beautifulsoup4 = "^4.12.3"
cryptography = "^43.0.0"
openpyxl = "^3.1.5"

# Required for CV module (always installed)
loguru = "^0.7.2"
supabase = "^2.10.0"
openai = "^1.0.0"
numpy = "^1.26.0"
opencv-python-headless = "^4.10.0"
pillow = "^10.0.0"

# Computer Vision Module Dependencies (optional - heavy packages for GPU/advanced features)
ultralytics = {version = "^8.3.0", optional = true}
easyocr = {version = "^1.7.2", optional = true}

[tool.poetry.extras]
cv-full = ["ultralytics", "easyocr"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
